Add inner join tests and fix hibernate problems

+++ testsuite-base/jpa/src/main/java/com/blazebit/persistence/testsuite/base/AbstractJpaPersistenceTest.java
@@ -301,7 +301,7 @@ public abstract class AbstractJpaPersistenceTest {
             factory.close();
         }
 
-        if (!databaseCleaner.supportsClearSchema()) {
+        if (databaseCleaner != null && !databaseCleaner.supportsClearSchema()) {
             dropSchema();
         }
     }
